requirements.txt
Flask
pytest
pyyaml
transformers
torch
flask-restx
python-dotenv
